--- logistics/src/main/java/com/alevel/deliverit/postal/network/Context.java
@@ -1,4 +1,0 @@
-package com.alevel.deliverit.postal.network;
-
-public class Context implements SendingContext {
-}
