+++ logistics/src/main/java/com/alevel/deliverit/postal/network/context/Context.java
@@ -1,0 +1,4 @@
+package com.alevel.deliverit.postal.network.context;
+
+public class Context implements SendingContext {
+}
