+++ logistics/src/main/java/com/alevel/deliverit/postal/network/dijkstra/Dijkstra.java
@@ -1,6 +1,7 @@
 package com.alevel.deliverit.postal.network.dijkstra;
 
 import com.alevel.deliverit.postal.network.*;
+import com.alevel.deliverit.postal.network.context.Context;
 import com.google.common.base.Preconditions;
 
 import java.util.*;
